% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octosql.R
\docType{methods}
\name{dbFetch,OctoSQLResult-method}
\alias{dbFetch,OctoSQLResult-method}
\title{Retrieve records from OctoSQL query}
\usage{
\S4method{dbFetch}{OctoSQLResult}(res, n = -1, ...)
}
\description{
Retrieve records from OctoSQL query
}
