% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octosql.R
\docType{methods}
\name{dbSendQuery,OctoSQLConnection-method}
\alias{dbSendQuery,OctoSQLConnection-method}
\title{Send a query to OctoSQL.}
\usage{
\S4method{dbSendQuery}{OctoSQLConnection}(conn, statement, ...)
}
\description{
Send a query to OctoSQL.
}
\examples{
# This is another good place to put examples
}
