% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octosql.R
\docType{methods}
\name{OctoSQL}
\alias{OctoSQL}
\alias{dbConnect,OctoSQLDriver-method}
\title{OctoSQL DBI wrapper}
\usage{
OctoSQL()

\S4method{dbConnect}{OctoSQLDriver}(drv, config, ...)
}
\arguments{
\item{drv}{An object created by \code{Athena()}}

\item{config}{path to config file}

\item{...}{Other options}
}
\description{
OctoSQL DBI wrapper

Authentication credentials are read from the DefaultAWSCredentialsProviderChain, which includes the .aws folder and
environment variables.
}
\examples{
\dontrun{
}
}
