% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octosql.R
\docType{methods}
\name{OctoSQL}
\alias{OctoSQL}
\alias{dbConnect,OctoSQLDriver-method}
\title{Athena DBI wrapper}
\usage{
OctoSQL()

\S4method{dbConnect}{OctoSQLDriver}(drv, config, ...)
}
\arguments{
\item{drv}{An object created by \code{Athena()}}

\item{...}{Other options}

\item{region}{the AWS region}

\item{S3OutputLocation}{S3 bucket where results will be saved to}

\item{Schema}{Athena schema to use}
}
\description{
Athena DBI wrapper

Authentication credentials are read from the DefaultAWSCredentialsProviderChain, which includes the .aws folder and
environment variables.
}
\examples{
\dontrun{
}
}
\seealso{
\href{http://docs.aws.amazon.com/athena/latest/ug/connect-with-jdbc.html#jdbc-options}{Athena Manual} for more connections options.
}
